refactor(auth): Move controllers to repositories
