+++ auth/Protos/authenticate.proto
@@ -7,6 +7,7 @@ package auth;
 service Auth {
     rpc Login(LoginRequest) returns (LoginResponse) {}
     rpc Register(RegisterRequest) returns (RegisterResponse) {}
+    rpc ValidateToken(ValidateTokenRequest) returns (ValidateTokenResponse) {}
 }
 
 message LoginRequest {
@@ -30,4 +31,13 @@ message RegisterResponse {
     string token = 1;
     string message = 2;
     int32 status_code = 3;
+}
+
+message ValidateTokenRequest {
+    string token = 1;
+}
+
+message ValidateTokenResponse {
+    string message = 1;
+    int32 status_code = 2;
 }
